Implemented all dashboard functionality, none of the monitor functionalities

+++ Managers/FileManager.py
@@ -1,6 +1,8 @@
 import json
 import os
 from Modules.error import BaseAppError
+import traceback  # traceback.print_exc() pritns the stack trace
+from typing import List
 
 class GamePathNotFoundError(BaseAppError):
     """Exception raised when the expected game directory is not found."""
@@ -38,6 +40,77 @@ class FileManager():
 
         self.recent_files_path = os.path.join(self.data_dir, "recent_files.json")
 
-        with open(self.recent_files_path, 'w') as file:  # instantiates the recent_files file
-            json.dump({}, file, indent=4)
+        if not os.path.exists(self.recent_files_path):
+            with open(self.recent_files_path, 'w') as file:  # instantiates the recent_files file
+                json.dump([], file, indent=4)
+
+    def file_selected(self, filepath: str) -> None:
+        self.save_to_recents(filepath)
+        self.create_file_slot(filepath)
+
+    def save_to_recents(self, filepath: str) -> None:
+        try:
+            recent_files = []
+            with open(self.recent_files_path, 'r') as file:
+                recent_files = json.load(file)
+
+            if filepath in recent_files:
+                recent_files.remove(filepath)
+
+            recent_files.insert(0, filepath)
+            recent_files = recent_files[:3]
+
+            with open(self.recent_files_path, 'w') as file:
+                json.dump(recent_files, file, indent=4)
+        except Exception as e:
+            traceback.print_exc()
+            print(e)
+
+    def get_recent_files(self) -> List:
+        try:
+            with open(self.recent_files_path, 'r') as file:
+                return json.load(file)
+        except Exception as e:
+            return []
+    
+    def create_file_slot(self, filepath: str) -> None:
+        new_filename_normal = os.path.basename(filepath).replace(".dat",".json")
+        new_filename_imp = os.path.basename(filepath).replace(".dat",".imp.json")
+
+        path_norm = os.path.join(self.data_dir,"files",new_filename_normal)
+        path_imp = os.path.join(self.data_dir,"files",new_filename_imp)
+
+        if not os.path.isfile(path_norm):
+            try:
+                with open(path_norm,'w') as f:
+                    json.dump({},f)
+            except Exception as e:
+                print(f"Exception occured while creating file slot: {e}")
+
+        if not os.path.isfile(path_imp):
+            try:
+                with open(path_imp,'w') as f:
+                    json.dump({},f)
+            except Exception as e:
+                print(f"Exception occured while creating file slot: {e}")
+
+    def load_change_logs(self, filepath: str) -> List:
+        filename = os.path.basename(filepath)
+
+        normal_changes_path = os.path.join(self.saves_dir, filename.replace(".dat",".json"))
+        imp_changes_path = os.path.join(self.saves_dir, filename.replace(".dat",".imp.json"))
+
+        logs=[]
+
+        with open(normal_changes_path,'r') as file:
+            logs.append(
+                json.load(file)
+            )
+        with open(imp_changes_path,'r') as file:
+            logs.append(
+                json.load(file)
+            )
+
+        return logs
+
